refactor: split project context into submodules
